+++ crawl-ref/source/notes.h
@@ -67,7 +67,7 @@ struct Note
 
     NOTE_TYPES type;
     int first, second;
-    long turn;
+    int turn;
     unsigned short packed_place;
 
     std::string name;
